requirements.txt
pandas>=1.1.0
psycopg2-binary>=2.9.1